Remove old TreeView and Provider and cleanup
