+++ app/(gcforms)/[locale]/(form administration)/form-builder/components/shared/right-panel/treeview/TreeView.tsx
@@ -1,87 +1,232 @@
-import React from "react";
-import { Button } from "@clientComponents/globals";
 import {
-  UncontrolledTreeEnvironment,
+  ControlledTreeEnvironment,
+  DraggingPosition,
+  DraggingPositionBetweenItems,
   Tree,
   TreeItem,
-  IndividualTreeViewState,
-  DraggingPosition,
-  DraggingPositionItem,
-  TreeDataProvider,
+  TreeItemIndex,
 } from "react-complex-tree";
-import "react-complex-tree/lib/style-modern.css";
-import { useTreeRef } from "./provider/TreeRefProvider";
 import { useGroupStore } from "./store/useGroupStore";
+import {
+  ForwardRefRenderFunction,
+  ReactElement,
+  forwardRef,
+  useImperativeHandle,
+  useState,
+} from "react";
+import { useTreeRef } from "./provider/TreeRefProvider";
 import { v4 as uuid } from "uuid";
-import { CustomStaticTreeDataProvider } from "./provider/CustomDataProvider";
+import { findParentGroup } from "./util/findParentGroup";
+import { Item } from "./Item";
 
-export const TreeView = ({
-  dataProvider,
-  viewState,
-  onFocusItem,
-}: {
-  dataProvider: CustomStaticTreeDataProvider;
-  viewState: IndividualTreeViewState;
-  onFocusItem: (item: TreeItem) => void;
-}) => {
-  const { environment, tree, wrapper } = useTreeRef();
+export interface TreeDataProviderProps {
+  children?: ReactElement;
+  addItem: (id: string) => void;
+  // addGroup: (id: string) => void;
+  updateItem: (id: string, value: string) => void;
+  // removeItem: (id: string) => void;
+  // openSection?: (id: string) => void;
+}
 
-  // This is only for testing purposes
-  // const getFocus = () => {
-  //   if (!environment || !environment.current) return "Fruit";
-  //   const state = environment.current.viewState;
-  //   return (state["tree-1"] && state["tree-1"].focusedItem) || "Fruit";
-  // };
+const ControlledTree: ForwardRefRenderFunction<unknown, TreeDataProviderProps> = (
+  { children },
+  ref
+) => {
+  // export const TreeView = () => {
+  const { getGroups, addGroup, setId, updateGroupName, updateGroup, updateElementTitle } =
+    useGroupStore((s) => {
+      return {
+        getGroups: s.getGroups,
+        addGroup: s.addGroup,
+        setId: s.setId,
+        updateGroupName: s.updateGroupName,
+        updateElementTitle: s.updateElementTitle,
+        updateGroup: s.updateGroup,
+      };
+    });
 
-  const updateElementTitle = useGroupStore((state) => state.updateElementTitle);
-  const updateGroupName = useGroupStore((state) => state.updateGroupName);
-  const updateGroup = useGroupStore((state) => state.updateGroup);
+  const { tree, environment } = useTreeRef();
+  const [focusedItem, setFocusedItem] = useState<TreeItemIndex | undefined>();
+  const [expandedItems, setExpandedItems] = useState<TreeItemIndex[]>([]);
+  const [selectedItems, setSelectedItems] = useState<TreeItemIndex[]>([]);
+
+  const addSection = () => {
+    const id = uuid();
+    addGroup(id, "New section");
+    setSelectedItems([id]);
+    setExpandedItems([id]);
+    setId(id);
+  };
+
+  useImperativeHandle(ref, () => ({
+    addItem: async (id: string) => {
+      const parent = findParentGroup(getGroups(), id);
+      setExpandedItems([parent?.index as TreeItemIndex]);
+      setSelectedItems([id]);
+    },
+    updateItem: (id: string) => {
+      const parent = findParentGroup(getGroups(), id);
+      setExpandedItems([parent?.index as TreeItemIndex]);
+      setSelectedItems([id]);
+    },
+  }));
 
   return (
-    <div className="relative mr-[1px]">
-      <div className="m-4">
-        <Button
-          onClick={() => {
-            const id = uuid();
-            wrapper?.current?.addGroup(id);
-          }}
-        >
-          Add section
-        </Button>
+    <ControlledTreeEnvironment
+      ref={environment}
+      items={getGroups()}
+      getItemTitle={(item) => item.data}
+      renderItem={({ title, arrow, context, children }) => {
+        return (
+          <Item title={title} arrow={arrow} context={context}>
+            {children}
+          </Item>
+        );
+      }}
+      renderItemTitle={({ title }) => <Item.Title title={title} />}
+      renderItemArrow={({ item, context }) => <Item.Arrow item={item} context={context} />}
+      renderLiveDescriptorContainer={() => null}
+      viewState={{
+        ["default"]: {
+          focusedItem,
+          expandedItems,
+          selectedItems,
+        },
+      }}
+      canDragAndDrop={true}
+      canReorderItems={true}
+      canDrag={(items: TreeItem[]) => {
+        return items.some((item) => {
+          return (
+            item.data !== "Start" &&
+            item.data !== "Introduction" &&
+            item.data !== "Policy" &&
+            item.data !== "End" &&
+            item.data !== "Confirmation"
+          );
+        });
+      }}
+      canDropAt={(items, target) => {
+        const folderItemsCount = items.filter((item) => item.isFolder).length;
 
-        <div>
-          <div>
-            <UncontrolledTreeEnvironment
-              ref={environment}
-              onFocusItem={onFocusItem}
-              onDrop={async (items: TreeItem[], target: DraggingPosition) => {
-                const { parentItem } = target as DraggingPositionItem;
-                const children = (await dataProvider.getTreeItem(parentItem)).children;
-                updateGroup(parentItem, children);
-              }}
-              onRenameItem={(item, name) => {
-                // Rename the group
-                item.isFolder && updateGroupName({ id: String(item.index), name });
+        // if any of the selected items is a folder, disallow dropping on a folder
+        if (folderItemsCount > 1) {
+          const { parentItem } = target as DraggingPositionBetweenItems;
+          if (items[0].isFolder && parentItem !== "root") {
+            return false;
+          }
+        }
 
-                // Rename the element
-                !item.isFolder &&
-                  updateElementTitle({
-                    id: Number(item.index),
-                    text: name,
-                  });
-              }}
-              canDragAndDrop={true}
-              canReorderItems={true}
-              dataProvider={dataProvider as unknown as TreeDataProvider}
-              getItemTitle={(item) => item.data || "New item"}
-              viewState={{ ...viewState }}
-              disableMultiselect={true}
-            >
-              <Tree treeId="tree-1" rootItem="root" treeLabel="Tree Example" ref={tree} />
-            </UncontrolledTreeEnvironment>
-          </div>
-        </div>
-      </div>
-    </div>
+        return true;
+      }}
+      onRenameItem={(item, name) => {
+        item.isFolder && updateGroupName({ id: String(item.index), name });
+
+        // Rename the element
+        !item.isFolder &&
+          updateElementTitle({
+            id: Number(item.index),
+            text: name,
+          });
+
+        setSelectedItems([item.index]);
+      }}
+      onDrop={async (items: TreeItem[], target: DraggingPosition) => {
+        let itemsPriorToInsertion = 0;
+
+        // current state of the tree
+        const currentItems = getGroups();
+
+        // Ids of the items being dragged
+        const itemsIndices = items.map((i) => i.index);
+
+        // Target parent and index
+        const { parentItem: targetParentIndex, childIndex: targetIndex } =
+          target as DraggingPositionBetweenItems;
+
+        // Loop over the items being dragged
+        items.forEach((item) => {
+          // Find the parent of the item being dragged
+          const originParent = findParentGroup(currentItems, String(item.index));
+
+          if (!originParent) {
+            throw Error(`Could not find parent of item "${item.index}"`);
+          }
+
+          if (!originParent.children) {
+            throw Error(
+              `Parent "${originParent.index}" of item "${item.index}" did not have any children`
+            );
+          }
+
+          if (target.targetType === "between-items" && target.parentItem === item.index) {
+            // Trying to drop inside itself
+            return;
+          }
+
+          // Origin index of the item being dragged
+          const originIndex = originParent.children.indexOf(String(item.index));
+
+          if (originIndex === -1) {
+            throw Error(`Item "${item.index}" not found in parent "${originParent.index}"`);
+          }
+
+          // Get the target parent
+          const targetParent = currentItems[targetParentIndex];
+
+          // Adjust the index if the item is being moved to a position after itself
+          const isOldItemPriorToNewItem =
+            ((targetParent.children ?? []).findIndex((child) => child === item.index) ?? Infinity) <
+            targetIndex;
+          itemsPriorToInsertion += isOldItemPriorToNewItem ? 1 : 0;
+
+          // Remove the item from the origin parent
+          originParent.children.splice(originIndex, 1);
+
+          // Update groups
+          updateGroup(originParent.index, originParent.children);
+        });
+
+        // Get the new state of the tree
+        const newItems = getGroups();
+
+        // Get the target parent in the new state
+        const targetParent = newItems[targetParentIndex];
+
+        // Initialize children if there are none
+        if (!targetParent.children) {
+          targetParent.children = [];
+        }
+
+        // Insert the items into the target parent
+        targetParent.children.splice(targetIndex - itemsPriorToInsertion, 0, ...itemsIndices);
+
+        // Update groups
+        updateGroup(targetParentIndex, targetParent.children);
+
+        // Set selected to trigger a re-render
+        setSelectedItems([targetParentIndex]);
+      }}
+      onFocusItem={(item) => {
+        setFocusedItem(item.index);
+        const parent = findParentGroup(getGroups(), String(item.index));
+        setId(item.isFolder ? String(item.index) : String(parent?.index));
+      }}
+      onExpandItem={(item) => setExpandedItems([...expandedItems, item.index])}
+      onCollapseItem={(item) =>
+        setExpandedItems(
+          expandedItems.filter((expandedItemIndex) => expandedItemIndex !== item.index)
+        )
+      }
+      onSelectItems={(items) => setSelectedItems(items)}
+    >
+      <Tree treeId="default" rootItem="root" treeLabel="Tree Example" ref={tree} />
+      <button className="ml-2 mt-2 rounded-md border border-slate-500 p-2" onClick={addSection}>
+        New section
+      </button>
+      <>{children}</>
+    </ControlledTreeEnvironment>
   );
 };
+
+export const TreeView = forwardRef(ControlledTree);
